+++ src/ReaderTask.ts
@@ -6,16 +6,13 @@ import { IO } from './IO'
 import { Monad2 } from './Monad'
 import { MonadTask2 } from './MonadTask'
 import { Monoid } from './Monoid'
-import { getSemigroup as getReaderSemigroup, Reader } from './Reader'
-import { getReaderM } from './ReaderT'
+import * as R from './Reader'
+import * as ReaderT from './ReaderT'
 import { Semigroup } from './Semigroup'
 import * as T from './Task'
 
+import Reader = R.Reader
 import Task = T.Task
-
-const MT =
-  /*#__PURE__*/
-  getReaderM(T.monadTask)
 
 /**
  * @since 2.3.0
@@ -52,14 +49,14 @@ export function run<R, A>(ma: ReaderTask<R, A>, r: R): Promise<A> {
  */
 export const fromTask: <R, A>(ma: Task<A>) => ReaderTask<R, A> =
   /*#__PURE__*/
-  (() => MT.fromM)()
+  R.of
 
 /**
  * @since 2.3.0
  */
 export const fromReader: <R, A = never>(ma: Reader<R, A>) => ReaderTask<R, A> =
   /*#__PURE__*/
-  (() => MT.fromReader)()
+  (() => ReaderT.fromReader(T.monadTask))()
 
 /**
  * @since 2.3.0
@@ -73,13 +70,13 @@ export function fromIO<R, A>(ma: IO<A>): ReaderTask<R, A> {
  */
 export const of: <R, A>(a: A) => ReaderTask<R, A> =
   /*#__PURE__*/
-  (() => MT.of)()
+  (() => ReaderT.of(T.monadTask))()
 
 /**
  * @since 2.3.0
  */
 export function getSemigroup<R, A>(S: Semigroup<A>): Semigroup<ReaderTask<R, A>> {
-  return getReaderSemigroup(T.getSemigroup<A>(S))
+  return R.getSemigroup(T.getSemigroup<A>(S))
 }
 
 /**
@@ -95,23 +92,14 @@ export function getMonoid<R, A>(M: Monoid<A>): Monoid<ReaderTask<R, A>> {
 /**
  * @since 2.3.0
  */
-export const ask: <R>() => ReaderTask<R, R> =
-  /*#__PURE__*/
-  (() => MT.ask)()
+export const ask: <R>() => ReaderTask<R, R> = () => T.of
 
 /**
  * @since 2.3.0
  */
 export const asks: <R, A = never>(f: (r: R) => A) => ReaderTask<R, A> =
   /*#__PURE__*/
-  (() => MT.asks)()
-
-/**
- * @since 2.3.0
- */
-export const local: <Q, R>(f: (f: Q) => R) => <A>(ma: ReaderTask<R, A>) => ReaderTask<Q, A> =
-  /*#__PURE__*/
-  (() => MT.local)()
+  (() => ReaderT.asks(T.monadTask))()
 
 /**
  * @since 2.4.0
@@ -152,7 +140,7 @@ export function chainTaskK<A, B>(f: (a: A) => Task<B>): <R>(ma: ReaderTask<R, A>
  */
 export const ap: <R, A>(fa: ReaderTask<R, A>) => <B>(fab: ReaderTask<R, (a: A) => B>) => ReaderTask<R, B> =
   /*#__PURE__*/
-  (() => MT.ap)()
+  (() => ReaderT.ap(T.monadTask))()
 
 /**
  * @since 2.3.0
@@ -179,7 +167,7 @@ export const apSecond = <R, B>(fb: ReaderTask<R, B>) => <A>(fa: ReaderTask<R, A>
  */
 export const chain: <A, R, B>(f: (a: A) => ReaderTask<R, B>) => (ma: ReaderTask<R, A>) => ReaderTask<R, B> =
   /*#__PURE__*/
-  (() => MT.chain)()
+  (() => ReaderT.chain(T.monadTask))()
 
 /**
  * @since 2.3.0
@@ -202,7 +190,7 @@ export const flatten: <R, A>(mma: ReaderTask<R, ReaderTask<R, A>>) => ReaderTask
  */
 export const map: <A, B>(f: (a: A) => B) => <R>(fa: ReaderTask<R, A>) => ReaderTask<R, B> =
   /*#__PURE__*/
-  (() => MT.map)()
+  (() => ReaderT.map(T.monadTask))()
 
 // -------------------------------------------------------------------------------------
 // instances